+++ CHANGELOG.md
@@ -18,6 +18,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v4.276.0 - 2022-11-10
+
+### Changed
+
+- 販売者のmakesOfferに適用取引期間を追加
+- PayActionを最適化
+- PayActionにlocationを追加
+- ISimpleOrderからprojectを削除
+- アクション検索条件拡張
+- アクションを最適化
+
 ## v4.275.0 - 2022-11-07
 
 ### Changed